requirements.txt
pendulum
python-jose
python-multipart
passlib
pydantic
aioredis==1.3.1
aiofiles==0.5.0
alembic==1.4.2
dynaconf==3.1.0
starlette==0.13.6
fastapi==0.61.1
FastAPI-SQLAlchemy==0.2.1
Flask==1.1.1
werkzeug==0.16.1
Flask-Admin==1.5.4
Flask-BasicAuth==0.2.0
Flask-SQLAlchemy==2.3.0
loguru==0.4.1
redis==3.5.3
requests==2.24.0
requests-async==0.6.2
sentry-sdk==0.16.5
SQLAlchemy==1.3.10
sqlalchemy-mixins==1.0
web3==5.0.2
websockets==8.1
python-socketio==4.6.0
uvicorn==0.12.1
psycopg2==2.8.6
aiojobs==0.2.2
python-engineio==3.13.2
nest_asyncio
